requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
flask-restx
pyjwt
python-dotenv